requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.22
celery==3.1.19
contextlib2==0.5.3
Django==1.10
django-braces==1.9.0
django-celery==3.1.17
django-cors-headers==1.1.0
django-countries==3.4.1
django-haystack==2.5.0
django-hitcount==1.2.2
django-model-utils==2.5
django-phonenumber-field==1.1.0
django-redis==4.3.0
django-taggit==0.20.2
djangorestframework==3.3.2
drf-nested-routers==0.11.1
easy-thumbnails==2.3
ecdsa==0.13
elasticsearch==1.7.0
elasticutils==0.10.2
Fabric==1.10.2
funcsigs==1.0.2
geoip2==2.2.0
ipaddress==1.0.16
jsonfield==1.0.3
kombu==3.0.33
maxminddb==1.2.0
meld3==1.0.2
mock==2.0.0
oauthlib==1.1.2
paramiko==1.16.0
pbr==1.10.0
phonenumberslite==7.4.3
Pillow==3.0.0
pkg-resources==0.0.0
psycopg2==2.6.1
pycrypto==2.6.1
PyJWT==1.4.0
python-memcached==1.57
python-openid==2.2.5
pytz==2015.7
raven==5.21.0
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
restless==2.0.1
six==1.10.0
stripe==1.36.0
unicode-slugify==0.1.3
Unidecode==0.4.18
urllib3==1.13.1
uWSGI==2.0.11.2
Whoosh==2.7.0
